Remove custom query handler altogether
It's easier to just handle Postgres errors in Waterline, which we can do now
since Waterline's only supported backend is Postgres.

The alternative would be to move all of the error handling into here but we
can't do that because Waterline performs validation and returns some errors
without ever hitting sails-postgresql, so defer all of the error handling
there.