Mark the database as dirty when a record is inserted
Currently we attempt to clear the database after every Shyp API test. However,
we only need to clear the database if we inserted records! If we can add logic
here and in DBConnection to check when the database is dirty, we can save a
database round trip between a large number of our tests.